backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
watchdog==3.0.0
docker==6.1.3
requests==2.31.0
python-dateutil==2.8.2
Werkzeug==2.3.7
click==8.1.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
six==1.16.0